requirements.txt
Django==2.2.2
django-crispy-forms==1.9.0
djangorestframework==3.10.1
mysqlclient==1.4.6
Pillow==6.1.0
pytz==2019.1
six==1.14.0
sqlparse==0.3.0
sweetify==2.0.0
